LoginModal перенесена в UserPanel
