+++ src/components/UserPanel/UserPanel.tsx
@@ -1,26 +1,33 @@
-import { FC } from 'react';
+import { FC, useCallback, useState } from 'react';
 
 import { useUser } from '@hooks/useUser';
 import { Button } from '@components/Button';
+import { LoginModal } from '@components/LoginModal';
 
-export type UserPanelProps = {
-  onLogout: () => void;
-  onLogin: () => void;
-};
+export const UserPanel: FC = (props) => {
+  const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
+  const { user, logout } = useUser();
 
-export const UserPanel: FC<UserPanelProps> = ({ onLogin, onLogout, ...props }) => {
-  const { user } = useUser();
+  function handleOpenLoginModal() {
+    setIsLoginModalOpen(true);
+  }
+
+  const handleCloseLoginModal = useCallback(() => {
+    setIsLoginModalOpen(false);
+  }, []);
 
   return (
     <div {...props}>
       {user.isLogin ? (
         <div>
           <h4>{user.username}</h4>
-          <Button onClick={onLogout}>Выйти</Button>
+          <Button onClick={logout}>Выйти</Button>
         </div>
       ) : (
-        <Button onClick={onLogin}>Войти</Button>
+        <Button onClick={handleOpenLoginModal}>Войти</Button>
       )}
+
+      {isLoginModalOpen && <LoginModal onClose={handleCloseLoginModal} />}
     </div>
   );
 };
